LayoutTests/media/video-playsinline-expected.txt

RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = false)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = false)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = false)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = false)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = true)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = true)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = true)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(true))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = true)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = false)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = false)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = false)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = false)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = true)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(true))
RUN(video.playsInline = true)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'true') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = true)
RUN(video.setAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
RUN(internals.settings.setAllowsInlineMediaPlaybackWithPlaysInlineAttribute(false))
RUN(internals.settings.setAllowsInlineMediaPlaybackWithWebKitPlaysInlineAttribute(false))
RUN(video.playsInline = true)
RUN(video.removeAttribute("webkit-playsinline", ""))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (video.webkitDisplayingFullscreen == 'false') OK
END OF TEST
